Move Unix domain sockets to their own type and only build them against Unix.
